Crate util_validation

A utility crate for securing code logic with unknown values and ensuring validation of data.

Example

use util_validation::{Validated, ValidationError, ValidateFrom};

pub struct RgbColor { red: u8, green: u8, blue: u8 }

impl ValidateFrom<&str> for RgbColor {
    fn validate(input: &str) -> Result<Validated<Self>, ValidationError> {
        // has to start with '#' and be followed by 6 hex digits
        if !input.starts_with('#') || input.len() != 7 {
            return Err(ValidationError::new("Invalid RGB color format"));
        }

        // Parse the hex values
        let red = u8::from_str_radix(&input[1..3], 16).map_err(|_| ValidationError::new("Invalid red component"))?;
        let green = u8::from_str_radix(&input[3..5], 16).map_err(|_| ValidationError::new("Invalid green component"))?;
        let blue = u8::from_str_radix(&input[5..7], 16).map_err(|_| ValidationError::new("Invalid blue component"))?;
        Ok(Validated(RgbColor { red, green, blue }))
    }
}

fn main() {
    let color_str = "#FF5733";
    let invalid_color_str = "FF5733";

    assert!(RgbColor::validate(color_str).is_ok());
    assert!(RgbColor::validate(invalid_color_str).is_err());
}

Features

util: Enables utility functions for validation, such as common validation patterns and helper functions.

uuid: Enables validation for UUID types. This allows you to validate UUIDs from strings and ensure they are in the correct format.

Semantics

Error messages should be short, concise and contextless. The context can be prepended along the stack of validation calls. The final error messages grows from the right to the left.

Error validating input: Component Name: Value must not be the SQL keyword `select`

For example the method not_sql_keyword does not need to know the field name, and speaks with words like “value” instead of specific field names. The context should be added by the caller.

Structs

Validated

A wrapper type that indicates that the inner value has been validated. This is used to ensure code logic that requires validated data can only be called with prior validation. Validated data is immutable, but the valid state can be dropped.

ValidationError

An error type for validation errors. This can be used to provide more context about why an input validation fails.

Traits

ValidateFrom

A trait for types that can be validated. The validate method should return a Validated wrapper around the original type

Functions

not_empty

Validates that the trimmed string value is not empty. This is a common validation pattern to assert a string is not just common whitespace.

not_sql_keyword

Validates that the string value is not a SQL keyword. This is a simple check against a hardcoded list of common SQL keywords. This is not exhaustive, but should cover the most common cases.

valid_sql_ident

Validates that the string value is a valid SQL identifier.

validate

A helper function to validate an input value using the ValidateFrom trait. This is a convenient way to perform validation without having to directly call the trait method.